% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab7SpaghettiBolognese.R
\docType{package}
\name{Lab7SpaghettiBolognese}
\alias{Lab7SpaghettiBolognese}
\alias{Lab7SpaghettiBolognese-package}
\title{Lab7SpaghettiBolognese: Linreg}
\description{
This package use API from googlemaps to download Longitude and Latitude of a address/destination so you can plot it!
}
\author{
Albin Vasterlund and Eric Herwin
}
