% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Do you want to make a Ridge Regression?}
\description{
This function make a "Reference Class" with class=ridgereg for a Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula for the model}

\item{\code{data}}{A object of class data.frame}

\item{\code{lambda}}{The lambda value}
}}


\examples{
data(iris)
ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris,0.5)$print()
ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris,0.5)$predict()
ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris,0.5)$coef()
}
