% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_leapForward.R
\name{train_leapForward}
\alias{train_leapForward}
\title{leapForward in a function}
\usage{
train_leapForward(formula, data)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data}
}
\description{
train(method=leapForward) dont work in Vignette
}
