% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Do you want to plot the mean delay time?}
\usage{
visualize_airport_delays()
}
\description{
Plats an interactive plot that displays the mean delay time for every airport
}
\section{Fields}{

\describe{
\item{\code{no}}{argument}
}}

\examples{
visualize_airport_delays()
}
