% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab7SpaghettiBolognese.R
\docType{package}
\name{Lab7SpaghettiBolognese}
\alias{Lab7SpaghettiBolognese}
\alias{Lab7SpaghettiBolognese-package}
\title{Lab7SpaghettiBolognese: Linreg}
\description{
Test
}
\author{
Albin Vasterlund and Eric Herwin
}
